pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentatlas"
version = "0.3.0"
description = "Shared browser interaction schema registry for AI agents. Reduces LLM token usage by 80-100% on known sites."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["ai", "agents", "browser-automation", "llm", "playwright"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "supabase>=2.6.0",
    "python-dotenv>=1.0.1",
    "playwright>=1.44.0",
    "openai>=1.109.1",
    "playwright-stealth>=1.0.6",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/agentatlas"
Issues = "https://github.com/yourusername/agentatlas/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["agentatlas*"]

[tool.pytest.ini_options]
python_files = ["test_*.py"]
testpaths = ["."]
